pyproject.toml
[project]
name = "deckdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8",
]

[project.scripts]
deckdown = "deckdown.cli:main"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest-cov>=7.0.0",
    "pytest>=8.4.2",
    "python-pptx>=1.0.2",
    "ruff>=0.13.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "dist",
    "build",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C90", "N", "S", "ANN"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "E501", "I001", "B905"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = false
strict_optional = true
show_error_codes = true
pretty = true
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true
mypy_path = ["src"]
files = ["src", "tests"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
